+++ cli/azd/cmd/logout.go
@@ -10,7 +10,6 @@ import (
 	"github.com/azure/azure-dev/cli/azd/cmd/actions"
 	"github.com/azure/azure-dev/cli/azd/internal"
 	"github.com/azure/azure-dev/cli/azd/pkg/auth"
-	"github.com/azure/azure-dev/cli/azd/pkg/commands"
 	"github.com/azure/azure-dev/cli/azd/pkg/output"
 	"github.com/spf13/cobra"
 )
@@ -20,9 +19,6 @@ func logoutCmdDesign(global *internal.GlobalCommandOptions) (*cobra.Command, *st
 		Use:   "logout",
 		Short: "Log out of Azure",
 		Long:  "Log out of Azure",
-		Annotations: map[string]string{
-			commands.RequireNoLoginAnnotation: "true",
-		},
 	}
 
 	return cmd, &struct{}{}