+++ cli/azd/cmd/templates.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/azure/azure-dev/cli/azd/cmd/actions"
 	"github.com/azure/azure-dev/cli/azd/internal"
-	"github.com/azure/azure-dev/cli/azd/pkg/commands"
 	"github.com/azure/azure-dev/cli/azd/pkg/output"
 	"github.com/azure/azure-dev/cli/azd/pkg/templates"
 	"github.com/spf13/cobra"
@@ -68,9 +67,6 @@ func templatesListCmdDesign(global *internal.GlobalCommandOptions) (*cobra.Comma
 		Use:     "list",
 		Short:   "List templates.",
 		Aliases: []string{"ls"},
-		Annotations: map[string]string{
-			commands.RequireNoLoginAnnotation: "true",
-		},
 	}
 
 	flags := &templatesListFlags{}
@@ -141,9 +137,6 @@ func templatesShowCmdDesign(rootOptions *internal.GlobalCommandOptions) (*cobra.
 	cmd := &cobra.Command{
 		Use:   "show <template>",
 		Short: "Show the template details.",
-		Annotations: map[string]string{
-			commands.RequireNoLoginAnnotation: "true",
-		},
 	}
 	output.AddOutputParam(cmd, []output.Format{output.JsonFormat, output.TableFormat}, output.TableFormat)