+++ cli/azd/pkg/commands/pipeline/github_provider.go
@@ -14,13 +14,13 @@ import (
 	"regexp"
 	"strings"
 
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/azure/azure-dev/cli/azd/pkg/azsdk"
 	"github.com/azure/azure-dev/cli/azd/pkg/convert"
 	"github.com/azure/azure-dev/cli/azd/pkg/environment"
 	githubRemote "github.com/azure/azure-dev/cli/azd/pkg/github"
 	"github.com/azure/azure-dev/cli/azd/pkg/graphsdk"
 	"github.com/azure/azure-dev/cli/azd/pkg/httputil"
-	"github.com/azure/azure-dev/cli/azd/pkg/identity"
 	"github.com/azure/azure-dev/cli/azd/pkg/infra/provisioning"
 	"github.com/azure/azure-dev/cli/azd/pkg/input"
 	"github.com/azure/azure-dev/cli/azd/pkg/output"
@@ -293,6 +293,13 @@ func notifyWhenGitHubActionsAreDisabled(
 // GitHubCiProvider implements a CiProvider using GitHub to manage CI pipelines as
 // GitHub actions.
 type GitHubCiProvider struct {
+	credential azcore.TokenCredential
+}
+
+func NewGitHubCiProvider(credential azcore.TokenCredential) *GitHubCiProvider {
+	return &GitHubCiProvider{
+		credential: credential,
+	}
 }
 
 // ***  subareaProvider implementation ******
@@ -404,7 +411,8 @@ func (p *GitHubCiProvider) configureConnection(
 	infraOptions provisioning.Options,
 	credentials json.RawMessage,
 	authType PipelineAuthType,
-	console input.Console) error {
+	console input.Console,
+) error {
 
 	repoSlug := repoDetails.owner + "/" + repoDetails.repoName
 	console.Message(ctx, fmt.Sprintf("Configuring repository %s.\n", output.WithHighLightFormat(repoSlug)))
@@ -425,7 +433,9 @@ func (p *GitHubCiProvider) configureConnection(
 	case AuthTypeClientCredentials:
 		authErr = p.configureClientCredentialsAuth(ctx, azdEnvironment, infraOptions, repoSlug, credentials, console)
 	default:
-		authErr = p.configureFederatedAuth(ctx, azdEnvironment, infraOptions, repoSlug, credentials, console)
+		authErr = p.configureFederatedAuth(
+			ctx, azdEnvironment, infraOptions, repoSlug, credentials, console, p.credential,
+		)
 	}
 
 	if authErr != nil {
@@ -525,6 +535,7 @@ func (p *GitHubCiProvider) configureFederatedAuth(
 	repoSlug string,
 	credentials json.RawMessage,
 	console input.Console,
+	credential azcore.TokenCredential,
 ) error {
 	ghCli := github.NewGitHubCli(ctx)
 
@@ -533,7 +544,7 @@ func (p *GitHubCiProvider) configureFederatedAuth(
 		return fmt.Errorf("failed unmarshalling azure credentials: %w", err)
 	}
 
-	err := applyFederatedCredentials(ctx, repoSlug, &azureCredentials, console)
+	err := applyFederatedCredentials(ctx, repoSlug, &azureCredentials, console, credential)
 	if err != nil {
 		return err
 	}
@@ -566,8 +577,9 @@ func applyFederatedCredentials(
 	repoSlug string,
 	azureCredentials *azcli.AzureCredentials,
 	console input.Console,
+	credential azcore.TokenCredential,
 ) error {
-	graphClient, err := createGraphClient(ctx)
+	graphClient, err := createGraphClient(ctx, credential)
 	if err != nil {
 		return err
 	}
@@ -815,12 +827,11 @@ func ensureFederatedCredential(
 	return nil
 }
 
-func createGraphClient(ctx context.Context) (*graphsdk.GraphClient, error) {
-	creds := identity.GetCredentials(ctx)
+func createGraphClient(ctx context.Context, credential azcore.TokenCredential) (*graphsdk.GraphClient, error) {
 	graphOptions := azsdk.
 		NewClientOptionsBuilder().
 		WithTransport(httputil.GetHttpClient(ctx)).
 		BuildCoreClientOptions()
 
-	return graphsdk.NewGraphClient(creds, graphOptions)
+	return graphsdk.NewGraphClient(credential, graphOptions)
 }