+++ cli/azd/pkg/infra/scope_test.go
@@ -17,6 +17,7 @@ import (
 	"github.com/azure/azure-dev/cli/azd/pkg/osutil"
 	"github.com/azure/azure-dev/cli/azd/pkg/tools/azcli"
 	"github.com/azure/azure-dev/cli/azd/test/mocks"
+	"github.com/azure/azure-dev/cli/azd/test/mocks/mockazcli"
 	"github.com/stretchr/testify/require"
 )
 
@@ -45,6 +46,7 @@ func TestScopeGetDeployment(t *testing.T) {
 
 	t.Run("SubscriptionScopeSuccess", func(t *testing.T) {
 		mockContext := mocks.NewMockContext(context.Background())
+		azCli := mockazcli.NewAzCliFromMockContext(mockContext)
 
 		subscriptionId := "SUBSCRIPTION_ID"
 		deploymentName := "DEPLOYMENT_NAME"
@@ -66,7 +68,7 @@ func TestScopeGetDeployment(t *testing.T) {
 			}, nil
 		})
 
-		scope := NewSubscriptionScope(*mockContext.Context, "eastus2", subscriptionId, deploymentName)
+		scope := NewSubscriptionScope(azCli, "eastus2", subscriptionId, deploymentName)
 
 		deployment, err := scope.GetDeployment(*mockContext.Context)
 		require.NoError(t, err)
@@ -77,6 +79,7 @@ func TestScopeGetDeployment(t *testing.T) {
 
 	t.Run("ResourceGroupScopeSuccess", func(t *testing.T) {
 		mockContext := mocks.NewMockContext(context.Background())
+		azCli := mockazcli.NewAzCliFromMockContext(mockContext)
 
 		subscriptionId := "SUBSCRIPTION_ID"
 		deploymentName := "DEPLOYMENT_NAME"
@@ -100,7 +103,7 @@ func TestScopeGetDeployment(t *testing.T) {
 			}, nil
 		})
 
-		scope := NewResourceGroupScope(*mockContext.Context, subscriptionId, resourceGroupName, deploymentName)
+		scope := NewResourceGroupScope(azCli, subscriptionId, resourceGroupName, deploymentName)
 
 		deployment, err := scope.GetDeployment(*mockContext.Context)
 		require.NoError(t, err)
@@ -118,6 +121,8 @@ func TestScopeDeploy(t *testing.T) {
 
 	t.Run("SubscriptionScopeSuccess", func(t *testing.T) {
 		mockContext := mocks.NewMockContext(context.Background())
+		azCli := mockazcli.NewAzCliFromMockContext(mockContext)
+
 		mockContext.HttpClient.When(func(request *http.Request) bool {
 			return request.Method == http.MethodPut && strings.Contains(
 				request.URL.Path,
@@ -133,7 +138,7 @@ func TestScopeDeploy(t *testing.T) {
 			}, nil
 		})
 
-		scope := NewSubscriptionScope(*mockContext.Context, "eastus2", "SUBSCRIPTION_ID", "DEPLOYMENT_NAME")
+		scope := NewSubscriptionScope(azCli, "eastus2", "SUBSCRIPTION_ID", "DEPLOYMENT_NAME")
 
 		armTemplate := azure.ArmTemplate(testArmTemplate)
 		err := scope.Deploy(*mockContext.Context, &armTemplate, parametersPath)
@@ -142,6 +147,8 @@ func TestScopeDeploy(t *testing.T) {
 
 	t.Run("ResourceGroupScopeSuccess", func(t *testing.T) {
 		mockContext := mocks.NewMockContext(context.Background())
+		azCli := mockazcli.NewAzCliFromMockContext(mockContext)
+
 		mockContext.HttpClient.When(func(request *http.Request) bool {
 			return request.Method == http.MethodPut && strings.Contains(
 				request.URL.Path,
@@ -158,7 +165,7 @@ func TestScopeDeploy(t *testing.T) {
 			}, nil
 		})
 
-		scope := NewResourceGroupScope(*mockContext.Context, "SUBSCRIPTION_ID", "RESOURCE_GROUP", "DEPLOYMENT_NAME")
+		scope := NewResourceGroupScope(azCli, "SUBSCRIPTION_ID", "RESOURCE_GROUP", "DEPLOYMENT_NAME")
 
 		armTemplate := azure.ArmTemplate(testArmTemplate)
 		err := scope.Deploy(*mockContext.Context, &armTemplate, parametersPath)
@@ -169,6 +176,8 @@ func TestScopeDeploy(t *testing.T) {
 func TestScopeGetResourceOperations(t *testing.T) {
 	t.Run("SubscriptionScopeSuccess", func(t *testing.T) {
 		mockContext := mocks.NewMockContext(context.Background())
+		azCli := mockazcli.NewAzCliFromMockContext(mockContext)
+
 		mockContext.HttpClient.When(func(request *http.Request) bool {
 			return request.Method == http.MethodGet && strings.Contains(
 				request.URL.Path,
@@ -184,7 +193,7 @@ func TestScopeGetResourceOperations(t *testing.T) {
 			}, nil
 		})
 
-		scope := NewSubscriptionScope(*mockContext.Context, "eastus2", "SUBSCRIPTION_ID", "DEPLOYMENT_NAME")
+		scope := NewSubscriptionScope(azCli, "eastus2", "SUBSCRIPTION_ID", "DEPLOYMENT_NAME")
 
 		operations, err := scope.GetResourceOperations(*mockContext.Context)
 		require.NoError(t, err)
@@ -193,6 +202,8 @@ func TestScopeGetResourceOperations(t *testing.T) {
 
 	t.Run("ResourceGroupScopeSuccess", func(t *testing.T) {
 		mockContext := mocks.NewMockContext(context.Background())
+		azCli := mockazcli.NewAzCliFromMockContext(mockContext)
+
 		mockContext.HttpClient.When(func(request *http.Request) bool {
 			return request.Method == http.MethodGet && strings.Contains(
 				request.URL.Path,
@@ -207,7 +218,7 @@ func TestScopeGetResourceOperations(t *testing.T) {
 				},
 			}, nil
 		})
-		scope := NewResourceGroupScope(*mockContext.Context, "SUBSCRIPTION_ID", "RESOURCE_GROUP", "DEPLOYMENT_NAME")
+		scope := NewResourceGroupScope(azCli, "SUBSCRIPTION_ID", "RESOURCE_GROUP", "DEPLOYMENT_NAME")
 
 		operations, err := scope.GetResourceOperations(*mockContext.Context)
 		require.NoError(t, err)