+++ cli/azd/pkg/project/framework_service_docker_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/azure/azure-dev/cli/azd/pkg/exec"
 	"github.com/azure/azure-dev/cli/azd/pkg/tools/docker"
 	"github.com/azure/azure-dev/cli/azd/test/mocks"
+	"github.com/azure/azure-dev/cli/azd/test/mocks/mockazcli"
 	"github.com/stretchr/testify/require"
 )
 
@@ -29,6 +30,8 @@ services:
 
 	env := environment.EphemeralWithValues("test-env", nil)
 	mockContext := mocks.NewMockContext(context.Background())
+	azCli := mockazcli.NewAzCliFromMockContext(mockContext)
+
 	mockContext.CommandRunner.When(func(args exec.RunArgs, command string) bool {
 		return strings.Contains(command, "docker build")
 	}).RespondFn(func(args exec.RunArgs) (exec.RunResult, error) {
@@ -50,7 +53,7 @@ services:
 
 	projectConfig, err := ParseProjectConfig(testProj, env)
 	require.NoError(t, err)
-	prj, err := projectConfig.GetProject(mockContext.Context, env)
+	prj, err := projectConfig.GetProject(*mockContext.Context, env, azCli)
 	require.NoError(t, err)
 
 	service := prj.Services[0]
@@ -101,6 +104,7 @@ services:
 
 	env := environment.EphemeralWithValues("test-env", nil)
 	mockContext := mocks.NewMockContext(context.Background())
+	azCli := mockazcli.NewAzCliFromMockContext(mockContext)
 
 	ran := false
 
@@ -128,7 +132,7 @@ services:
 	projectConfig, err := ParseProjectConfig(testProj, env)
 	require.NoError(t, err)
 
-	prj, err := projectConfig.GetProject(mockContext.Context, env)
+	prj, err := projectConfig.GetProject(*mockContext.Context, env, azCli)
 	require.NoError(t, err)
 
 	service := prj.Services[0]